Reduce ring buffer size to 2 instead of 50
The XISL library is not deallocating some memory on its side in between
runs. We actually process the frames faster than they come in, so we
don't even need the ring buffer space, and reducing its size seems to
greatly reduce the "memory leak".

Maybe a better long-term solution is to not re-connect to the detectors
in between runs. If we do this, then we don't have to allocate new frame
buffer space each time. However, this probably requires too many changes
to make before the final experiment night tonight.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>