Fix issue with invalid dark arrays
These numpy arrays were referring to memory in the ring buffer, which
gets modified after a certain number of frames have been acquired. This
was resulting in an issue where the median would sometimes contain ambient
signal. So we *must* make a copy of the numpy arrays.

Thankfully, all of the data (dark, ambient, and shot) is valid, and we can
just re-compute the dark median and "*_ds" files to fix this.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>