Stop the acquisition after median subtraction
It doesn't take that long to do the median subtraction, and
for some reason, stopping the acquisition takes like 11 seconds. This
was keeping users from seeing the preview until stopping the acquisition
finished. Just stop the acquisition after indicating that we are
finished, so that users can see the preview much faster.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>